requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
bleach==6.0.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.1
defusedxml==0.7.1
Django==4.2
django-autoslug==1.9.8
django-birthday==0.1.4
django-cities-light==3.9.2
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==22.1
django-image-uploader-widget==0.2.5
django-jazzmin==2.6.0
django-model-utils==4.3.1
django-phonenumber-field==7.0.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-api-logger==1.1.12
drf-spectacular==0.26.1
drf-yasg==1.21.5
exceptiongroup==1.1.0
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
loguru==0.6.0
MarkupSafe==2.1.2
model-bakery==1.10.1
mypy==1.2.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==22.0
phonenumbers==8.13.10
phonenumberslite==8.13.10
Pillow==9.3.0
pluggy==1.0.0
progressbar2==4.2.0
promise==2.3
prompt-toolkit==3.0.36
psycopg2==2.9.6
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
pytest==7.2.1
python-utils==3.4.5
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.1
sqlparse==0.4.3
swapper==1.3.0
text-unidecode==1.3
tomli==2.0.1
tqdm==4.64.1
typing_extensions==4.5.0
tzdata==2023.3
Unidecode==1.3.6
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
win32-setctime==1.1.0